// src/app/user.interface.ts
export interface User {
      car_capacity: number
      car_color: string //"#baacac"
      car_color_name_en: string //"ang"
      car_color_name_hy: string //"hay"
      car_color_name_ru: string // "rus"
      car_model: string//"opel"
      car_number: string// "12321"
      first_name: string// "joni"
      image?: string// ""
      last_name: string //"brjoni"
      located_city_id: number //1
      main_city_id: number // 1
      phone_number: number // "555666"
      viber_id?: null
}